+++ R/utils-pnp-4x6.R
@@ -63,6 +63,8 @@ draw_4x6_tiles <- function(cfg, i_suit) {
     popViewport()
 
     df$piece_side = "tile_back"
+    # Rotate tile backs to partially hide direction of face if tile back are not fully symmetric
+    df$angle <- 90 * ((df$suit + df$rank) %% 4)
     grid.newpage()
     pushViewport(vp)
     pmap_piece(df, cfg = cfg, default.units = "npc")